reorganize api auth module
