+++ api/src/auth/checks/__tests__/check-permission.test.js
@@ -1,11 +1,11 @@
 import { dbNameFromFile } from 'arango-tools'
-import { ensureDatabase as ensure } from '../../testUtilities'
+import { ensureDatabase as ensure } from '../../../testUtilities'
 import { setupI18n } from '@lingui/core'
 
 import { checkPermission } from '..'
-import englishMessages from '../../locale/en/messages'
-import frenchMessages from '../../locale/fr/messages'
-import dbschema from '../../../database.json'
+import englishMessages from '../../../locale/en/messages'
+import frenchMessages from '../../../locale/fr/messages'
+import dbschema from '../../../../database.json'
 
 const { DB_PASS: rootPass, DB_URL: url } = process.env
