+++ api/src/auth/checks/check-domain-permission.js
