+++ api/src/auth/utils/verify-jwt.js
